update new structure
