+++ .pre-commit-config.yaml
@@ -18,8 +18,8 @@ repos:
     rev: v2.3.0
     hooks:
       - id: add-trailing-comma
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.982  # Use the sha / tag you want to point at
-    hooks:
-      - id: mypy
-        args: [--no-strict-optional, --ignore-missing-imports]
+#  - repo: https://github.com/pre-commit/mirrors-mypy
+#    rev: v0.982  # Use the sha / tag you want to point at
+#    hooks:
+#      - id: mypy
+#        args: [--no-strict-optional, --ignore-missing-imports]
