+++ pyproject.toml
@@ -11,6 +11,7 @@ fastapi = "^0.85.1"
 SQLAlchemy = { version = "^1.4.42", extras = ["asyncio"] }
 Werkzeug = "^2.2.2"
 pydantic = {version = "^1.10.2", extras = ["dotenv"]}
+inflection = "^0.5.1"
 
 [tool.poetry.dev-dependencies]
 uvicorn = "^0.19.0"
