+++ src/fastapi_oauth/common/encoding.py
@@ -1,9 +1,9 @@
 import base64
 import json
-import struct
+from typing import Optional
 
 
-def to_bytes(x, charset='utf-8', errors='strict'):
+def to_bytes(x, charset='utf-8', errors='strict') -> Optional[bytes]:
     if x is None:
         return None
     if isinstance(x, bytes):
@@ -15,7 +15,7 @@ def to_bytes(x, charset='utf-8', errors='strict'):
     return bytes(x)
 
 
-def to_unicode(x, charset='utf-8', errors='strict'):
+def to_unicode(x, charset='utf-8', errors='strict') -> str:
     if x is None or isinstance(x, str):
         return x
     if isinstance(x, bytes):
@@ -23,7 +23,7 @@ def to_unicode(x, charset='utf-8', errors='strict'):
     return str(x)
 
 
-def to_native(x, encoding='ascii'):
+def to_native(x, encoding='ascii') -> str:
     if isinstance(x, str):
         return x
     return x.decode(encoding)
@@ -44,23 +44,3 @@ def urlsafe_b64decode(s):
 
 def urlsafe_b64encode(s):
     return base64.urlsafe_b64encode(s).rstrip(b'=')
-
-
-def base64_to_int(s):
-    data = urlsafe_b64decode(to_bytes(s, charset='ascii'))
-    buf = struct.unpack('%sB' % len(data), data)
-    return int(''.join(['%02x' % byte for byte in buf]), 16)
-
-
-def int_to_base64(num):
-    if num < 0:
-        raise ValueError('Must be a positive integer')
-
-    s = num.to_bytes((num.bit_length() + 7) // 8, 'big', signed=False)
-    return to_unicode(urlsafe_b64encode(s))
-
-
-def json_b64encode(text):
-    if isinstance(text, dict):
-        text = json_dumps(text)
-    return urlsafe_b64encode(to_bytes(text))
