+++ src/fastapi_oauth/common/setting.py
