+++ src/fastapi_oauth/common/types.py
@@ -2,8 +2,8 @@ from typing import Any, Callable, Coroutine, Dict, Optional, Protocol
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from .models import ClientMixin, TokenMixin
-from .wrappers import OAuth2Request
+from fastapi_oauth.rfc6749.models import ClientMixin, TokenMixin
+from fastapi_oauth.rfc6749.wrappers import OAuth2Request
 
 QueryClientFn = Callable[[str, AsyncSession], Coroutine[Any, Any, ClientMixin]]
 QueryTokenFn = Callable[[str, str, AsyncSession], Coroutine[Any, Any, TokenMixin]]
